+++ cmem_plugin_reason/utils.py
@@ -20,6 +20,12 @@ from defusedxml import minidom
 
 from . import __path__
 
+with (Path(__path__[0]) / "reason_doc.md").open("r") as f:
+    REASON_DOC = f.read()
+
+with (Path(__path__[0]) / "validate_doc.md").open("r") as f:
+    VALIDATE_DOC = f.read()
+
 REASONERS = OrderedDict(
     {
         "elk": "ELK",