+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 
  - defined input and output schema
+ - Reason: parameters to import the result graph in the ontology graph and to import the ontology graph in the result graph
+ - Validate: parameter to enable/disable entity output
 
 ### Fixed